@@ -1,6 +1,4 @@
-from __future__ import absolute_import
-
-import mock
+from unittest import mock
 import platform
 import sys
 
