@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 import os
 import sys
 
@@ -7,7 +5,7 @@ import pytest
 from qtpy import PYSIDE2, PYSIDE6
 
 @pytest.mark.skipif(
-    sys.platform != "win32" or os.environ['USE_CONDA'] == 'Yes' or PYSIDE6,
+    sys.platform != "win32" or os.environ.get('USE_CONDA', 'Yes') == 'Yes' or PYSIDE6,
     reason="Only available in Qt5 bindings > 5.9 (only available with pip in the current CI setup) and Windows platform")
 def test_qtwinextras():
     """Test the qtpy.QtWinExtras namespace"""
